Split up execution logic
