+++ Stormpath.SDK/Stormpath.SDK/Impl/Linq/CollectionResourceQueryProvider.cs
@@ -13,18 +13,18 @@ namespace Stormpath.SDK.Impl.Linq
 {
     internal sealed class CollectionResourceQueryProvider<TResult> : IAsyncQueryProvider<TResult>, IQueryProvider
     {
-        private readonly string collectionHref;
-        private readonly IInternalDataStore dataStore;
-        private readonly CollectionResourceExecutor executor;
+        private CollectionResourceExecutor<TResult> executor;
 
-        public CollectionResourceQueryProvider(string collectionHref, IInternalDataStore dataStore)
+        public CollectionResourceQueryProvider(CollectionResourceExecutor<TResult> executor)
         {
-            this.collectionHref = collectionHref;
-            this.dataStore = dataStore;
-            this.executor = new CollectionResourceExecutor(dataStore);
+            this.executor = executor;
         }
 
-        public string CollectionHref => this.collectionHref;
+        public CollectionResourceExecutor<TResult> Executor
+        {
+            get { return this.executor; }
+            set { this.executor = value; }
+        }
 
         // Collection-returning standard query operators call this method.
         IAsyncQueryable<TResult> IAsyncQueryProvider<TResult>.CreateQuery(Expression expression)
@@ -36,12 +36,6 @@ namespace Stormpath.SDK.Impl.Linq
         public IQueryable<T> CreateQuery<T>(Expression expression)
             => new CollectionResourceQueryable<T>(this as IAsyncQueryProvider<T>, expression);
 
-        public Task<CollectionResponsePage<T>> ExecuteCollectionAsync<T>(string href, CancellationToken cancellationToken)
-            => this.executor.ExecuteCollectionAsync<T>(href, cancellationToken);
-
-        public CollectionResponsePage<T> ExecuteCollection<T>(string href)
-            => this.executor.ExecuteCollection<T>(href);
-
         public object Execute(Expression expression)
         {
             throw new NotImplementedException();
@@ -51,7 +45,7 @@ namespace Stormpath.SDK.Impl.Linq
         {
             bool isEnumerable = typeof(T).Name == "IEnumerable`1";
 
-            var compiledModel = 
+            //var compiledModel = 
 
             //return (TResult)SimpleExecutor.Execute(expression, IsEnumerable);
             throw new NotImplementedException();
